requirements.txt
beautifulsoup4
requests
argparse
python-dotenv
fbchat
